[Task] - Vector and String add small changes. - Add new type for file metadata. - Add more converter. - Ignore script before new file system is ready for run script. - Add more getcomponent chuice. - Fix converter bugs. - Refactor some class and fucntion name. - Add FileDatabase ability to find file by GUID and support more file types. - World manager now can save and load game world. - Add file browser GUI. - Ignore gameplay main until file browser is ready to work.

+++ CrossPlatform/Converter/Vector2YamlConverter.h
@@ -23,8 +23,8 @@ namespace YAML {
             if (!node.IsSequence() || node.size() != 3) {
                 return false;
             }
-            rhs.x = node[0].as<double>();
-            rhs.y = node[1].as<double>();
+            rhs.x = node[0].as<float>();
+            rhs.y = node[1].as<float>();
             return true;
         }
     };
